further refactored main.py by extracting routes into individual modules
